requirements.txt
beautifulsoup4==4.13.4
faiss-cpu==1.11.0
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
matplotlib==3.10.3
plotly==6.2.0
python-dotenv==1.1.1
requests==2.32.4
sentence-transformers==5.0.0
streamlit==1.46.1
tiktoken==0.9.0